This adds 3 commands:
* `tabs:pin-tab`
* `tabs:unpin-tab`
* `tabs:toggle-pin-tab`

This adds the `Toggle pin tab` command to both palette commands and as
`File` sub menu.

This adds the `Pin Tab` or `Unpin Tab` context menu to tabs (depending
on current status)

Pinned tabs titles are bolden and a `lock` icon is shown instead of the
`x` close-icon (unless tab is hovered).

Pinned tabs are **not** closed on following events:

* close-all-tabs
* close-other-tabs
* close-tabs-to-right
* close-saved-tabs

_Note:_ pinned tabs are still closed on the regular `close-tab` command