Read pane/item directly off tab element instead of looking up via DOM
Determining a tab’s pane or item via the DOM is unnecessary because we
store this state directly on each tab. It was leading to exceptions in
cases where a tab’s pane was being removed as a side-effect of dragging
because the tab didn’t have a parent element. This happens when the
‘core.removeEmptyPanes’ config setting is enabled.

Signed-off-by: Max Brunsfeld <maxbrunsfeld@github.com>
Signed-off-by: Antonio Scandurra <as-cii@github.com>